pyproject.toml

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hassl"
version = "0.1.0"
description = "HASSL: Home Assistant Simple Scripting Language"
authors = [{ name="adanowitz" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["lark-parser", "jinja2", "pyyaml"]

[project.scripts]
hasslc = "hassl.cli:main"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["hassl*"]
